Update README coverage of config.example.toml file
The existing coverage in the Examples section opened up with
a reference to `config.toml` while the example commands used
the `config.example.toml` starter file.

Those examples retain the `config.example.toml` file usage,
but we expand the coverage of where the `config.toml` file
should be located for automatic loading support.