Add support for SRV record protocol shortcuts
Allow using specific keywords as shortcuts in place of more
complex query strings. For exmaple, this allows specifying
the `msdcs` keyword along with the `example.com` query
string in place of `SRV` or `srv` query type along with
`_ldap._tcp.dc._msdcs.example.com` as the query string.

If specifying a SRV protocol via a supported keyword,
the SRV query type is assumed if not explicitly specified.
This is intended as a convenience for the user.

Documentation and example config file changes to
illustrate use of the new functionality have been applied.

The concurrency implementation has been refactored to
better support the new SRV protocol functionality and to
make future changes a little easier to make.

refs GH-139