Implement multi-line summary output support
Set multi-line summary output display as the default with
the option to switch back to the previous display format.

- Config file setting support ("persistent" display setting)
- Flag support (one-off override)

Doc updates have been applied to reflect the new default
output summary display format, including the addition of
one example that uses the previous display format. This example
attempts to illustrate what the new multi-line output format
is attempting to solve.

refs GH-146